chore: moved type defination to interfaces dir
