+++ packages/elements-react/lib/components/Filters/src/utils/createQuery.ts
@@ -1,4 +1,4 @@
-import { Condition } from "../types";
+import type { Condition } from "../interfaces/src/FilterInterface";
 
 const createQuery = (conditions: Condition[]) => {
   const qFrag = conditions.map(
